Fix for Status Effects and Healing being cancelled. I had added a player data save from ASB but missed a later commit that fixed resulting bugs. This adds the fix they made. https://github.com/AirSkyBoat/AirSkyBoat/commit/a391bb286f4904742c4e6785569a661955733b35

+++ src/map/status_effect_container.cpp
@@ -1695,7 +1695,7 @@ void CStatusEffectContainer::LoadStatusEffects()
  *                                                                       *
  ************************************************************************/
 
-void CStatusEffectContainer::SaveStatusEffects(bool logout)
+void CStatusEffectContainer::SaveStatusEffects(bool logout, bool skipRemove)
 {
     // Print entity name and bail out if entity isn't a player.
     if (m_POwner->objtype != TYPE_PC)
@@ -1709,7 +1709,7 @@ void CStatusEffectContainer::SaveStatusEffects(bool logout)
 
     for (CStatusEffect* PStatusEffect : m_StatusEffectSet)
     {
-        if ((logout && PStatusEffect->GetFlag() & EFFECTFLAG_LOGOUT) || (!logout && PStatusEffect->GetFlag() & EFFECTFLAG_ON_ZONE))
+        if (!skipRemove && ((logout && PStatusEffect->GetFlag() & EFFECTFLAG_LOGOUT) || (!logout && PStatusEffect->GetFlag() & EFFECTFLAG_ON_ZONE)))
         {
             RemoveStatusEffect(PStatusEffect, true);
             continue;
@@ -1769,7 +1769,10 @@ void CStatusEffectContainer::SaveStatusEffects(bool logout)
                        std::chrono::duration_cast<std::chrono::seconds>(PStatusEffect->GetStartTime().time_since_epoch()).count());
         }
     }
-    DeleteStatusEffects();
+    if (!skipRemove)
+    {
+        DeleteStatusEffects();
+    }
 }
 
 /************************************************************************
